rename files to match commands
