+++ src/commands/enable-security-alerts.js
