+++ src/commands/enable-security-fixes.js
